DORM-11 refactor(login): transfer registration component logic to login component, delete registration-page
