+++ src/app/contants/router-constants.ts
@@ -1,6 +1,6 @@
 export const RouterConstants = {
   main: { path: 'main', name: 'Main Page' },
-  register: { path: 'register', name: 'Register' },
+  login: { path: 'login', name: 'Login' },
   detail: { path: 'detail/:id', name: 'Detail' },
   initial: { path: '', name: 'Initial Page' },
   notFound: { path: '**', name: 'Not Found' }
