--- src/app/login/registration-page/registration-page.component.html
@@ -1,28 +1,0 @@
-<form class="register wrapper" (ngSubmit)="onSubmit()" #signUpForm="ngForm">
-  <h3>{{pageName}}</h3>
-
-<div class="register__container">
-  <app-google-input class="register__input-field" [placeholder]="'Login...'" [label]="'login'" (valueChanged)="changedLogin($event)"></app-google-input>
-  <app-google-input class="register__input-field" [placeholder]="'Password...'" [label]="'password'" (valueChanged)="changedPassword($event)"></app-google-input>
-  <app-google-input class="register__input-field" [placeholder]="'Repeat password...'" [label]="'password'" (valueChanged)="changedRepeatPassword($event)"></app-google-input>
-</div>
-
-
-
-  <p
-    [hidden]="isConfirm"
-  >Passwords do not match!</p>
-
-  <div class="buttons">
-    <button type="submit" *ngIf="pageState">LogIn</button>
-    <button *ngIf="!pageState" (click)="goToLogin()">Back</button>
-
-    <button *ngIf="pageState" (click)="goToRegister()">Register</button>
-    <button type="submit"
-      [disabled]="signUpForm.form.invalid"
-      *ngIf="!pageState"
-    >Register</button>
-  </div>
-</form>
-
-
