--- src/app/login/registration-page/registration-page.component.scss
@@ -1,44 +1,0 @@
-:host {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-}
-
-h3 {
-  margin: 0 0 20px 0;
-}
-
-.register {
-  width: 320px;
-  margin-top: 7vh;
-
-  &__container {
-    height: 180px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-  }
-}
-
-.buttons {
-  display: flex;
-  justify-content: space-around;
-}
-
-label {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  margin: 5px 10px;
-  font-size: 15px;
-}
-
-button {
-  margin-top: 10px;
-}
-
-p {
-  color: red;
-  text-align: center;
-}
