+++ src/app/shared/google-input/google-input.component.html
@@ -1,7 +1,8 @@
 <div class="input-container">
-    <app-google-wrapper [label]="label" [placeholder]="placeholder" [hasInnerFieldValue]="input.value"> 
+    <app-google-wrapper [label]="label" [placeholder]="placeholder" [hasInnerFieldValue]="input.value">
       <input #input
-                   class="input-container__element" value="{{ value }}"
-                   (input)="onValueChanged(input.value)"/>
+        [name]="name" [type]="type"
+        class="input-container__element" value="{{ value }}"
+        (input)="onValueChanged(input.value)"/>
     </app-google-wrapper>
 </div>
