+++ src/app/shared/google-input/google-input.component.scss
@@ -1,7 +1,6 @@
 @import '../../../styles/variables.scss';
 
 .input-container {
-  margin-left: 16px;
 
   &__element {
      margin: 0;
