+++ src/app/shared/google-wrapper/google-wrapper.component.scss
@@ -3,7 +3,8 @@
 .google-wrapper {
     position: relative;
     width: fit-content;
-  
+    margin: 8px 0;
+
     &__label {
       position: absolute;
       left: 8px;
@@ -12,13 +13,13 @@
       z-index: 1;
       transition: all 0.2s;
     }
-  
+
     .border {
       font-size: $font-s;
       line-height: 1.33;
       transform: translateY(-8px);
       color: $color-grey;
-  
+
       &::before {
         //100% width of the parent element + 8px white space (4px on left and right side)
         width: calc(100% + 6px);
@@ -31,7 +32,7 @@
         left: -4px;
       }
     }
-  
+
     .placeholder {
       font-size: $font-m;
       line-height: 1.2;
