+++ src/styles.scss
@@ -7,6 +7,11 @@ body {
 }
 
 button {
+  height: 35px;
+  min-width: 90px;
+  padding: 0 7px;
+  border: none;
+  font-size: 14px;
   cursor: pointer;
 
   &:disabled {
